Module raftstore::coprocessor

Re-exports

pub use self::config::Config;

pub use self::config::ConsistencyCheckMethod;

pub use self::dispatcher::BoxAdminObserver;

pub use self::dispatcher::BoxApplySnapshotObserver;

pub use self::dispatcher::BoxCmdObserver;

pub use self::dispatcher::BoxConsistencyCheckObserver;

pub use self::dispatcher::BoxQueryObserver;

pub use self::dispatcher::BoxRegionChangeObserver;

pub use self::dispatcher::BoxRoleObserver;

pub use self::dispatcher::BoxSplitCheckObserver;

pub use self::dispatcher::CoprocessorHost;

pub use self::dispatcher::Registry;

pub use self::region_info_accessor::Callback as RegionInfoCallback;

pub use self::region_info_accessor::RegionCollector;

pub use self::region_info_accessor::RegionInfo;

pub use self::region_info_accessor::RegionInfoAccessor;

pub use self::region_info_accessor::RegionInfoProvider;

pub use self::region_info_accessor::SeekRegionCallback;

Modules

config

consistency_check

dispatcher

error

metrics

region_info_accessor

split_check

split_observer

Structs

Cmd

CmdBatch

HalfCheckObserver

KeyEntry

KeysCheckObserver

ObserverContext
Context of observer.

RawConsistencyCheckObserver

SizeCheckObserver

SplitCheckerHost

TableCheckObserver

Enums

Error

RegionChangeEvent

Traits

AdminObserver

ApplySnapshotObserver

CmdObserver

ConsistencyCheckObserver

Coprocessor
Coprocessor is used to provide a convenient way to inject code to KV processing.

QueryObserver

RegionChangeObserver

RoleObserver

SplitCheckObserver

SplitChecker
SplitChecker is invoked during a split check scan, and decides to use which keys to split a region.

Functions

get_region_approximate_keys
Get the approximate number of keys in the range.

get_region_approximate_keys_cf

get_region_approximate_middle
Get region approximate middle key based on default and write cf size.

get_region_approximate_size
Get the approximate size of the range.

get_region_approximate_size_cf

Type Definitions

Result