Struct raftstore::coprocessor::region_info_accessor::RegionCollector

pub struct RegionCollector {
    regions: HashMap<u64, RegionInfo>,
    region_ranges: BTreeMap<Vec<u8>, u64>,
}

RegionCollector is the place where we hold all region information we collected, and the underlying runner of RegionInfoAccessor. It listens on events sent by the RegionEventListener and keeps information of all regions. Role of each region are also tracked.

Fields
regions: HashMap<u64, RegionInfo>region_ranges: BTreeMap<Vec<u8>, u64>
Implementations

impl RegionCollector

pub fn new() -> Self

pub fn create_region(&mut self, region: Region, role: StateRole)

fn update_region(&mut self, region: Region)

fn handle_create_region(&mut self, region: Region, role: StateRole)

fn handle_update_region(&mut self, region: Region, role: StateRole)

fn handle_destroy_region(&mut self, region: Region)

fn handle_role_change(&mut self, region: Region, new_role: StateRole)

fn is_region_epoch_stale(
&self,
region_to_check: &Region,
current: &Region
) -> bool

Determines whether region_to_check's epoch is stale compared to current's epoch

fn check_region_range(
&mut self,
region: &Region,
clear_regions_in_range: bool
) -> bool

For all regions whose range overlaps with the given region or region_id is the same as region's, checks whether the given region's epoch is not older than theirs.

Returns false if the given region is stale, which means, at least one region above has newer epoch. If the given region is not stale, all other regions in the collection that overlaps with the given region must be stale. Returns true in this case, and if clear_regions_in_range is true, those out-of-date regions will be removed from the collection.

pub fn handle_seek_region(
&self,
from_key: Vec<u8>,
callback: SeekRegionCallback
)

pub fn handle_find_region_by_id(
&self,
region_id: u64,
callback: Callback<Option<RegionInfo>>
)

pub fn handle_get_regions_in_range(
&self,
start_key: Vec<u8>,
end_key: Vec<u8>,
callback: Callback<Vec<Region>>
)

fn handle_raftstore_event(&mut self, event: RaftStoreEvent)