Trait raftstore::coprocessor::region_info_accessor::RegionInfoProvider

pub trait RegionInfoProvider: Send + Clone + 'static {
    fn seek_region(
        &self, 
        _from: &[u8], 
        _callback: SeekRegionCallback
    ) -> Result<()> { ... }

    fn find_region_by_id(
        &self, 
        _reigon_id: u64, 
        _callback: Callback<Option<RegionInfo>>
    ) -> Result<()> { ... }

    fn get_regions_in_range(
        &self, 
        _start_key: &[u8], 
        _end_key: &[u8]
    ) -> Result<Vec<Region>> { ... }
}

Provided methods

fn seek_region(&self, _from: &[u8], _callback: SeekRegionCallback) -> Result<()>

Get a iterator of regions that contains from or have keys larger than from, and invoke the callback to process the result.

fn find_region_by_id(
&self,
_reigon_id: u64,
_callback: Callback<Option<RegionInfo>>
) -> Result<()>

fn get_regions_in_range(
&self,
_start_key: &[u8],
_end_key: &[u8]
) -> Result<Vec<Region>>

Loading content...