Trait raftstore::coprocessor::RoleObserver

pub trait RoleObserver: Coprocessor {
    fn on_role_change(&self, _: &mut ObserverContext, _: StateRole) { ... }
}

Provided methods

fn on_role_change(&self, _: &mut ObserverContext, _: StateRole)

Hook to call when role of a peer changes.

Please note that, this hook is not called at realtime. There maybe a situation that the hook is not called yet, however the role of some peers have changed.

Loading content...