Trait raftstore::coprocessor::SplitChecker

pub trait SplitChecker<E> {
    fn split_keys(&mut self) -> Vec<Vec<u8>>;

    fn policy(&self) -> CheckPolicy;

    fn on_kv(&mut self, _: &mut ObserverContext<'_>, _: &KeyEntry) -> bool { ... }

    fn approximate_split_keys(
        &mut self, 
        _: &Region, 
        _: &E
    ) -> Result<Vec<Vec<u8>>> { ... }
}

SplitChecker is invoked during a split check scan, and decides to use which keys to split a region.

Required methods

fn split_keys(&mut self) -> Vec<Vec<u8>>

Get the desired split keys.

fn policy(&self) -> CheckPolicy

Get split policy.

Loading content...
Provided methods

fn on_kv(&mut self, _: &mut ObserverContext<'_>, _: &KeyEntry) -> bool

Hook to call for every kv scanned during split.

Return true to abort scan early.

fn approximate_split_keys(&mut self, _: &Region, _: &E) -> Result<Vec<Vec<u8>>>

Get approximate split keys without scan.

Loading content...