Trait raftstore::router::RaftStoreRouter

pub trait RaftStoreRouter<EK>: StoreRouter<EK> + ProposalRouter<EK::Snapshot> + CasualRouter<EK> + Send + Clone where
    EK: KvEngine, {
    fn send_raft_msg(&self, msg: RaftMessage) -> RaftStoreResult<()>;

    fn significant_send(
        &self, 
        region_id: u64, 
        msg: SignificantMsg<EK::Snapshot>
    ) -> RaftStoreResult<()>;

    fn broadcast_normal(&self, msg_gen: impl FnMut() -> PeerMsg<EK>);

    fn send_casual_msg(
        &self, 
        region_id: u64, 
        msg: CasualMessage<EK>
    ) -> RaftStoreResult<()> { ... }

    fn send_store_msg(&self, msg: StoreMsg<EK>) -> RaftStoreResult<()> { ... }

    fn send_command(
        &self, 
        req: RaftCmdRequest, 
        cb: Callback<EK::Snapshot>
    ) -> RaftStoreResult<()> { ... }

    fn report_unreachable(
        &self, 
        region_id: u64, 
        to_peer_id: u64
    ) -> RaftStoreResult<()> { ... }

    fn report_snapshot_status(
        &self, 
        region_id: u64, 
        to_peer_id: u64, 
        status: SnapshotStatus
    ) -> RaftStoreResult<()> { ... }

    fn broadcast_unreachable(&self, store_id: u64) { ... }

    fn report_resolved(&self, store_id: u64, group_id: u64) { ... }
}

Routes messages to the raftstore.

Required methods

fn send_raft_msg(&self, msg: RaftMessage) -> RaftStoreResult<()>

Sends RaftMessage to local store.

fn significant_send(
&self,
region_id: u64,
msg: SignificantMsg<EK::Snapshot>
) -> RaftStoreResult<()>

Sends a significant message. We should guarantee that the message can't be dropped.

fn broadcast_normal(&self, msg_gen: impl FnMut() -> PeerMsg<EK>)

Broadcast a message generated by msg_gen to all Raft groups.

Loading content...
Provided methods

fn send_casual_msg(
&self,
region_id: u64,
msg: CasualMessage<EK>
) -> RaftStoreResult<()>

Send a casual message to the given region.

fn send_store_msg(&self, msg: StoreMsg<EK>) -> RaftStoreResult<()>

Send a store message to the backend raft batch system.

fn send_command(
&self,
req: RaftCmdRequest,
cb: Callback<EK::Snapshot>
) -> RaftStoreResult<()>

Sends RaftCmdRequest to local store.

fn report_unreachable(
&self,
region_id: u64,
to_peer_id: u64
) -> RaftStoreResult<()>

Reports the peer being unreachable to the Region.

fn report_snapshot_status(
&self,
region_id: u64,
to_peer_id: u64,
status: SnapshotStatus
) -> RaftStoreResult<()>

Reports the sending snapshot status to the peer of the Region.

fn broadcast_unreachable(&self, store_id: u64)

Broadcast an StoreUnreachable event to all Raft groups.

fn report_resolved(&self, store_id: u64, group_id: u64)

Report a StoreResolved event to all Raft groups.

Loading content...