Module raftstore::store::fsm

Generally peers are state machines that represent a replica of a region, and store is also a special state machine that handles all requests across stores. They are mixed for now, will be separated in the future.

Re-exports

pub use self::apply::create_apply_batch_system;

pub use self::apply::Apply;

pub use self::apply::ApplyBatchSystem;

pub use self::apply::ApplyMetrics;

pub use self::apply::ApplyRes;

pub use self::apply::ApplyRouter;

pub use self::apply::Builder as ApplyPollerBuilder;

pub use self::apply::CatchUpLogs;

pub use self::apply::ChangeCmd;

pub use self::apply::ChangePeer;

pub use self::apply::ExecResult;

pub use self::apply::GenSnapTask;

pub use self::apply::Msg as ApplyTask;

pub use self::apply::Notifier as ApplyNotifier;

pub use self::apply::ObserveID;

pub use self::apply::Proposal;

pub use self::apply::Registration;

pub use self::apply::TaskRes as ApplyTaskRes;

pub use self::store::create_raft_batch_system;

pub use self::store::RaftBatchSystem;

pub use self::store::RaftPollerBuilder;

pub use self::store::RaftRouter;

pub use self::store::StoreInfo;

pub use self::store::StoreMeta;

Modules

apply

metrics

peer

store

Structs

DestroyPeerJob

PeerFsm