Struct raftstore::store::fsm::peer::PeerFsm

pub struct PeerFsm<EK, ER> where
    EK: KvEngine,
    ER: RaftEngine,  {
    pub peer: Peer<EK, ER>,
    tick_registry: PeerTicks,
    missing_ticks: usize,
    group_state: GroupState,
    stopped: bool,
    has_ready: bool,
    early_apply: bool,
    mailbox: Option<BasicMailbox<PeerFsm<EK, ER>>>,
    pub receiver: Receiver<PeerMsg<EK>>,
    skip_split_count: usize,
    batch_req_builder: BatchRaftCmdRequestBuilder<EK>,
}

Fields
peer: Peer<EK, ER>tick_registry: PeerTicks

A registry for all scheduled ticks. This can avoid scheduling ticks twice accidentally.

missing_ticks: usize

Ticks for speed up campaign in chaos state.

Followers will keep ticking in Idle mode to measure how many ticks have been skipped. Once it becomes chaos, those skipped ticks will be ticked so that it can campaign quickly instead of waiting an election timeout.

This will be reset to 0 once it receives any messages from leader.

group_state: GroupStatestopped: boolhas_ready: boolearly_apply: boolmailbox: Option<BasicMailbox<PeerFsm<EK, ER>>>receiver: Receiver<PeerMsg<EK>>skip_split_count: usize

when snapshot is generating or sending, skip split check at most REGION_SPLIT_SKIT_MAX_COUNT times.

batch_req_builder: BatchRaftCmdRequestBuilder<EK>
Implementations

impl<EK, ER> PeerFsm<EK, ER> where
EK: KvEngine,
ER: RaftEngine,

pub fn create(
store_id: u64,
cfg: &Config,
sched: Scheduler<RegionTask<EK::Snapshot>>,
engines: Engines<EK, ER>,
region: &Region
) -> Result<(LooseBoundedSender<PeerMsg<EK>>, Box<PeerFsm<EK, ER>>)>

pub fn replicate(
store_id: u64,
cfg: &Config,
sched: Scheduler<RegionTask<EK::Snapshot>>,
engines: Engines<EK, ER>,
region_id: u64,
peer: Peer
) -> Result<(LooseBoundedSender<PeerMsg<EK>>, Box<PeerFsm<EK, ER>>)>

pub fn region_id(&self) -> u64

pub fn get_peer(&self) -> &Peer<EK, ER>

pub fn peer_id(&self) -> u64

pub fn stop(&mut self)

pub fn set_pending_merge_state(&mut self, state: MergeState)

pub fn schedule_applying_snapshot(&mut self)