Struct raftstore::store::fsm::store::PollContext

pub struct PollContext<EK, ER, T, C: 'static> where
    EK: KvEngine,
    ER: RaftEngine,  {
    pub cfg: Config,
    pub store: Store,
    pub pd_scheduler: FutureScheduler<PdTask<EK>>,
    pub consistency_check_scheduler: Scheduler<ConsistencyCheckTask<EK::Snapshot>>,
    pub split_check_scheduler: Scheduler<SplitCheckTask>,
    pub cleanup_scheduler: Scheduler<CleanupTask>,
    pub raftlog_gc_scheduler: Scheduler<RaftlogGcTask<ER>>,
    pub region_scheduler: Scheduler<RegionTask<EK::Snapshot>>,
    pub apply_router: ApplyRouter<EK>,
    pub router: RaftRouter<EK, ER>,
    pub importer: Arc<SSTImporter>,
    pub store_meta: Arc<Mutex<StoreMeta>>,
    pub pending_create_peers: Arc<Mutex<HashMap<u64, (u64, bool)>>>,
    pub poller_handle: Handle,
    pub raft_metrics: RaftMetrics,
    pub snap_mgr: SnapManager,
    pub applying_snap_count: Arc<AtomicUsize>,
    pub coprocessor_host: CoprocessorHost<EK>,
    pub timer: SteadyTimer,
    pub trans: T,
    pub pd_client: Arc<C>,
    pub global_replication_state: Arc<Mutex<GlobalReplicationState>>,
    pub global_stat: GlobalStoreStat,
    pub store_stat: LocalStoreStat,
    pub engines: Engines<EK, ER>,
    pub kv_wb: EK::WriteBatch,
    pub raft_wb: ER::LogBatch,
    pub pending_count: usize,
    pub sync_log: bool,
    pub has_ready: bool,
    pub ready_res: Vec<(Ready, InvokeContext)>,
    pub need_flush_trans: bool,
    pub current_time: Option<Timespec>,
    pub perf_context_statistics: PerfContextStatistics,
    pub node_start_time: Option<TiInstant>,
}

Fields
cfg: Configstore: Storepd_scheduler: FutureScheduler<PdTask<EK>>consistency_check_scheduler: Scheduler<ConsistencyCheckTask<EK::Snapshot>>split_check_scheduler: Scheduler<SplitCheckTask>cleanup_scheduler: Scheduler<CleanupTask>raftlog_gc_scheduler: Scheduler<RaftlogGcTask<ER>>region_scheduler: Scheduler<RegionTask<EK::Snapshot>>apply_router: ApplyRouter<EK>router: RaftRouter<EK, ER>importer: Arc<SSTImporter>store_meta: Arc<Mutex<StoreMeta>>pending_create_peers: Arc<Mutex<HashMap<u64, (u64, bool)>>>

region_id -> (peer_id, is_splitting) Used for handling race between splitting and creating new peer. An uninitialized peer can be replaced to the one from splitting iff they are exactly the same peer.

WARNING: To avoid deadlock, if you want to use store_meta and pending_create_peers together, the lock sequence MUST BE:

lock the store_meta.

lock the pending_create_peers.

poller_handle: Handleraft_metrics: RaftMetricssnap_mgr: SnapManagerapplying_snap_count: Arc<AtomicUsize>coprocessor_host: CoprocessorHost<EK>timer: SteadyTimertrans: Tpd_client: Arc<C>global_replication_state: Arc<Mutex<GlobalReplicationState>>global_stat: GlobalStoreStatstore_stat: LocalStoreStatengines: Engines<EK, ER>kv_wb: EK::WriteBatchraft_wb: ER::LogBatchpending_count: usizesync_log: boolhas_ready: boolready_res: Vec<(Ready, InvokeContext)>need_flush_trans: boolcurrent_time: Option<Timespec>perf_context_statistics: PerfContextStatisticsnode_start_time: Option<TiInstant>
Implementations

impl<EK, ER, T, C> PollContext<EK, ER, T, C> where
EK: KvEngine,
ER: RaftEngine,

pub fn store_id(&self) -> u64

pub fn is_hibernate_timeout(&mut self) -> bool

Timeout is calculated from TiKV start, the node should not become hibernated if it still within the hibernate timeout, see https://github.com/tikv/tikv/issues/7747

impl<EK, ER, T: Transport, C> PollContext<EK, ER, T, C> where
EK: KvEngine,
ER: RaftEngine,

fn schedule_store_tick(&self, tick: StoreTick, timeout: Duration)

pub fn handle_stale_msg(
&mut self,
msg: &RaftMessage,
cur_epoch: RegionEpoch,
need_gc: bool,
target_region: Option<Region>
)