Struct raftstore::store::fsm::store::StoreFsmDelegate

struct StoreFsmDelegate<'a, EK: KvEngine + 'static, ER: RaftEngine + 'static, T: 'static> {
    fsm: &'a mut StoreFsm<EK>,
    ctx: &'a mut PollContext<EK, ER, T>,
}

Fields
fsm: &'a mut StoreFsm<EK>ctx: &'a mut PollContext<EK, ER, T>
Implementations

impl<'a, EK: KvEngine + 'static, ER: RaftEngine + 'static, T: Transport> StoreFsmDelegate<'a, EK, ER, T>

fn on_tick(&mut self, tick: StoreTick)

fn handle_msgs(&mut self, msgs: &mut Vec<StoreMsg<EK>>)

fn start(&mut self, store: Store)

impl<'a, EK: KvEngine, ER: RaftEngine, T: Transport> StoreFsmDelegate<'a, EK, ER, T>

fn check_msg(&mut self, msg: &RaftMessage) -> Result<CheckMsgStatus>

Checks if the message is targeting a stale peer.

fn on_raft_message(&mut self, msg: RaftMessage) -> Result<()>

fn maybe_create_peer(
&mut self,
region_id: u64,
msg: &RaftMessage,
is_local_first: bool
) -> Result<bool>

If target peer doesn't exist, create it.

return false to indicate that target peer is in invalid state or doesn't exist and can't be created.

fn maybe_create_peer_internal(
&mut self,
region_id: u64,
msg: &RaftMessage,
is_local_first: bool
) -> Result<bool>

fn on_compaction_finished(&mut self, event: EK::CompactedEvent)

fn register_compact_check_tick(&self)

fn on_compact_check_tick(&mut self)

fn store_heartbeat_pd(&mut self)

fn on_pd_store_heartbeat_tick(&mut self)

fn handle_snap_mgr_gc(&mut self) -> Result<()>

fn on_snap_mgr_gc(&mut self)

fn on_compact_lock_cf(&mut self)

fn register_pd_store_heartbeat_tick(&self)

fn register_snap_mgr_gc_tick(&self)

fn register_compact_lock_cf_tick(&self)

impl<'a, EK: KvEngine, ER: RaftEngine, T: Transport> StoreFsmDelegate<'a, EK, ER, T>

fn on_validate_sst_result(&mut self, ssts: Vec<SstMeta>)

fn on_cleanup_import_sst(&mut self) -> Result<()>

fn register_consistency_check_tick(&mut self)

fn on_consistency_check_tick(&mut self)

fn on_cleanup_import_sst_tick(&mut self)

fn register_cleanup_import_sst_tick(&self)

fn clear_region_size_in_range(&mut self, start_key: &[u8], end_key: &[u8])

fn on_store_unreachable(&mut self, store_id: u64)

fn on_update_replication_mode(&mut self, status: ReplicationStatus)

fn register_raft_engine_purge_tick(&self)

fn on_raft_engine_purge_tick(&self)

fn on_check_leader(
&self,
leaders: Vec<LeaderInfo>,
cb: Box<dyn FnOnce(Vec<u64>) + Send>
)