Module raftstore::store::msg

Structs

PeerTicks

RaftCommand
Raft command is the command that is expected to be proposed by the leader of the target raft group.

ReadResponse

WriteResponse

Enums

Callback
Variants of callbacks for Msg.

CasualMessage
Message that will be sent to a peer.

MergeResultKind

PeerMsg
Message that can be sent to a peer.

SignificantMsg
Some significant messages sent to raftstore. Raftstore will dispatch these messages to Raft groups to update some important internal status.

StoreMsg

StoreTick

Type Definitions

ReadCallback

WriteCallback