Module raftstore::store::peer

Structs

CheckTickResult

CmdEpochChecker

ConsistencyState
ConsistencyState is used for consistency check.

Peer

PeerStat
Statistics about raft peer.

ProposalContext
A bitmap contains some useful flags when dealing with eraftpb::Entry.

ProposalQueue

ProposedAdminCmd

Enums

RequestPolicy
RequestPolicy decides how we handle a request.

StaleState
The returned states of the peer after checking whether it is stale

Constants

MIN_BCAST_WAKE_UP_INTERVAL

REGION_READ_PROGRESS_CAP

SHRINK_CACHE_CAPACITY

Traits

AbstractPeer
A poor version of Peer to avoid port generic variables everywhere.

RequestInspector
RequestInspector makes RequestPolicy for requests.

Functions

get_sync_log_from_request

get_transfer_leader_cmd

is_request_urgent
We enable follower lazy commit to get a better performance. But it may not be appropriate for some requests. This function checks whether the request should be committed on all followers as soon as possible.

make_transfer_leader_response