Module raftstore::store::region_snapshot

Structs

RegionIterator
RegionIterator wrap a rocksdb iterator and only allow it to iterate in the region. It behaves as if underlying db only contains one region.

RegionSnapshot
Snapshot of a region.

Functions

handle_check_key_in_region_error

update_lower_bound

update_upper_bound