Module raftstore::store::snap::snap_io

Structs

BuildStatistics

Traits

StaleDetector
Used to check a procedure is stale or not.

Functions

apply_plain_cf_file
Apply the given snapshot file into a column family. callback will be invoked after each batch of key value pairs written to db.

apply_sst_cf_file

build_plain_cf_file
Build a snapshot file for the given column family in plain format. If there are no key-value pairs fetched, no files will be created at path, otherwise the file will be created and synchronized.

build_sst_cf_file
Build a snapshot file for the given column family in sst format. If there are no key-value pairs fetched, no files will be created at path, otherwise the file will be created and synchronized.

create_sst_file_writer

get_decrypter_reader