Function raftstore::store::snap::snap_io::apply_plain_cf_file

pub fn apply_plain_cf_file<E, F>(
    path: &str, 
    key_mgr: Option<&Arc<DataKeyManager>>, 
    stale_detector: &impl StaleDetector, 
    db: &E, 
    cf: &str, 
    batch_size: usize, 
    callback: F
) -> Result<(), Error> where
    E: KvEngine,
    F: for<'r> FnMut(&'r [(Vec<u8>, Vec<u8>)]), 

Apply the given snapshot file into a column family. callback will be invoked after each batch of key value pairs written to db.