Struct raftstore::store::LocalReader

pub struct LocalReader<C, E> where
    C: ProposalRouter<E::Snapshot>,
    E: KvEngine,  {
    store_id: Cell<Option<u64>>,
    store_meta: Arc<Mutex<StoreMeta>>,
    kv_engine: E,
    metrics: ReadMetrics,
    delegates: HashMap<u64, Option<ReadDelegate>>,
    snap_cache: Option<Arc<E::Snapshot>>,
    cache_read_id: ThreadReadId,
    router: C,
}

Fields
store_id: Cell<Option<u64>>store_meta: Arc<Mutex<StoreMeta>>kv_engine: Emetrics: ReadMetricsdelegates: HashMap<u64, Option<ReadDelegate>>snap_cache: Option<Arc<E::Snapshot>>cache_read_id: ThreadReadIdrouter: C
Implementations

impl<C, E> LocalReader<C, E> where
C: ProposalRouter<E::Snapshot>,
E: KvEngine,

pub fn new(kv_engine: E, store_meta: Arc<Mutex<StoreMeta>>, router: C) -> Self

fn redirect(&mut self, cmd: RaftCommand<E::Snapshot>)

fn pre_propose_raft_command(
&mut self,
req: &RaftCmdRequest
) -> Result<Option<ReadDelegate>>

pub fn propose_raft_command(
&mut self,
read_id: Option<ThreadReadId>,
req: RaftCmdRequest,
cb: Callback<E::Snapshot>
)

pub fn read(
&mut self,
read_id: Option<ThreadReadId>,
req: RaftCmdRequest,
cb: Callback<E::Snapshot>
)

If read requests are received at the same RPC request, we can create one snapshot for all of them and check whether the time when the snapshot was created is in lease. We use ThreadReadId to figure out whether this RaftCommand comes from the same RPC request with the last RaftCommand which left a snapshot cached in LocalReader. ThreadReadId is composed by thread_id and a thread_local incremental sequence.

pub fn release_snapshot_cache(&mut self)