Struct raftstore::store::Peer

pub struct Peer<EK, ER> where
    EK: KvEngine,
    ER: RaftEngine,  {
    region_id: u64,
    pub tag: String,
    pub peer: Peer,
    pub raft_group: RawNode<PeerStorage<EK, ER>>,
    peer_cache: RefCell<HashMap<u64, Peer>>,
    pub peer_heartbeats: HashMap<u64, Instant>,
    proposals: ProposalQueue<EK::Snapshot>,
    leader_missing_time: Option<Instant>,
    leader_lease: Lease,
    pending_reads: ReadIndexQueue<EK::Snapshot>,
    pub leader_unreachable: bool,
    pub should_wake_up: bool,
    pub pending_remove: bool,
    pending_messages: Vec<Message>,
    pub peers_start_pending_time: Vec<(u64, Instant)>,
    down_peer_ids: Vec<u64>,
    pub size_diff_hint: u64,
    delete_keys_hint: u64,
    pub compaction_declined_bytes: u64,
    pub approximate_size: Option<u64>,
    pub approximate_keys: Option<u64>,
    pub consistency_state: ConsistencyState,
    pub pending_request_snapshot_count: Arc<AtomicUsize>,
    pub last_applying_idx: u64,
    pub last_compacted_idx: u64,
    last_urgent_proposal_idx: u64,
    last_committed_split_idx: u64,
    pub raft_log_size_hint: u64,
    last_proposed_prepare_merge_idx: u64,
    last_committed_prepare_merge_idx: u64,
    pub pending_merge_state: Option<MergeState>,
    pub want_rollback_merge_peers: HashSet<u64>,
    pub catch_up_logs: Option<CatchUpLogs>,
    pub peer_stat: PeerStat,
    pub bcast_wake_up_time: Option<UtilInstant>,
    pub replication_mode_version: u64,
    pub dr_auto_sync_state: DrAutoSyncState,
    pub replication_sync: bool,
    pub check_stale_conf_ver: u64,
    pub check_stale_peers: Vec<Peer>,
    pub local_first_replicate: bool,
    pub txn_extra_op: Arc<AtomicCell<TxnExtraOp>>,
    cmd_epoch_checker: CmdEpochChecker<EK::Snapshot>,
}

Fields
region_id: u64

The ID of the Region which this Peer belongs to.

tag: String

Peer_tag, "[region <region_id>] <peer_id>"

peer: Peer

The Peer meta information.

raft_group: RawNode<PeerStorage<EK, ER>>

The Raft state machine of this Peer.

peer_cache: RefCell<HashMap<u64, Peer>>

The cache of meta information for Region's other Peers.

peer_heartbeats: HashMap<u64, Instant>

Record the last instant of each peer's heartbeat response.

proposals: ProposalQueue<EK::Snapshot>leader_missing_time: Option<Instant>leader_lease: Leasepending_reads: ReadIndexQueue<EK::Snapshot>leader_unreachable: bool

If it fails to send messages to leader.

should_wake_up: bool

Indicates whether the peer should be woken up.

pending_remove: bool

Whether this peer is destroyed asynchronously. If it's true,

when merging, its data in storeMeta will be removed early by the target peer.

all read requests must be rejected.

pending_messages: Vec<Message>

If a snapshot is being applied asynchronously, messages should not be sent.

peers_start_pending_time: Vec<(u64, Instant)>

Record the instants of peers being added into the configuration. Remove them after they are not pending any more.

down_peer_ids: Vec<u64>

A inaccurate cache about which peer is marked as down.

size_diff_hint: u64

An inaccurate difference in region size since last reset. It is used to decide whether split check is needed.

delete_keys_hint: u64

The count of deleted keys since last reset.

compaction_declined_bytes: u64

An inaccurate difference in region size after compaction. It is used to trigger check split to update approximate size and keys after space reclamation of deleted entries.

approximate_size: Option<u64>

Approximate size of the region.

approximate_keys: Option<u64>

Approximate keys of the region.

consistency_state: ConsistencyState

The state for consistency check.

pending_request_snapshot_count: Arc<AtomicUsize>

The counter records pending snapshot requests.

last_applying_idx: u64

The index of last scheduled committed raft log.

last_compacted_idx: u64

The index of last compacted raft log. It is used for the next compact log task.

last_urgent_proposal_idx: u64

The index of the latest urgent proposal index.

last_committed_split_idx: u64

The index of the latest committed split command.

raft_log_size_hint: u64

Approximate size of logs that is applied but not compacted yet.

last_proposed_prepare_merge_idx: u64

The index of the latest proposed prepare merge command.

last_committed_prepare_merge_idx: u64

The index of the latest committed prepare merge command.

pending_merge_state: Option<MergeState>

The merge related state. It indicates this Peer is in merging.

want_rollback_merge_peers: HashSet<u64>

The rollback merge proposal can be proposed only when the number of peers is greater than the majority of all peers. There are more details in the annotation above test_node_merge_write_data_to_source_region_after_merging The peers who want to rollback merge

catch_up_logs: Option<CatchUpLogs>

source region is catching up logs for merge

peer_stat: PeerStat

Write Statistics for PD to schedule hot spot.

bcast_wake_up_time: Option<UtilInstant>

Time of the last attempt to wake up inactive leader.

replication_mode_version: u64

Current replication mode version.

dr_auto_sync_state: DrAutoSyncState

The required replication state at current version.

replication_sync: bool

A flag that caches sync state. It's set to true when required replication state is reached for current region.

check_stale_conf_ver: u64

The known newest conf version and its corresponding peer list Send to these peers to check whether itself is stale.

check_stale_peers: Vec<Peer>local_first_replicate: bool

Whether this peer is created by replication and is the first one of this region on local store.

txn_extra_op: Arc<AtomicCell<TxnExtraOp>>cmd_epoch_checker: CmdEpochChecker<EK::Snapshot>

Check whether this proposal can be proposed based on its epoch

Implementations

impl<EK, ER> Peer<EK, ER> where
EK: KvEngine,
ER: RaftEngine,

pub fn new(
store_id: u64,
cfg: &Config,
sched: Scheduler<RegionTask<EK::Snapshot>>,
engines: Engines<EK, ER>,
region: &Region,
peer: Peer
) -> Result<Peer<EK, ER>>

pub fn init_replication_mode(&mut self, state: &mut GlobalReplicationState)

Sets commit group to the peer.

pub fn switch_replication_mode(&mut self, state: &Mutex<GlobalReplicationState>)

Updates replication mode.

pub fn activate<T, C>(&self, ctx: &PollContext<EK, ER, T, C>)

Register self to apply_scheduler so that the peer is then usable. Also trigger RegionChangeEvent::Create here.

fn next_proposal_index(&self) -> u64

pub fn get_index_term(&self, idx: u64) -> u64

pub fn maybe_append_merge_entries(
&mut self,
merge: &CommitMergeRequest
) -> Option<u64>

pub fn maybe_destroy<T, C>(
&mut self,
ctx: &PollContext<EK, ER, T, C>
) -> Option<DestroyPeerJob>

Tries to destroy itself. Returns a job (if needed) to do more cleaning tasks.

pub fn destroy<T, C>(
&mut self,
ctx: &PollContext<EK, ER, T, C>,
keep_data: bool
) -> Result<()>

Does the real destroy task which includes:

Set the region to tombstone;

Clear data;

Notify all pending requests.

pub fn is_initialized(&self) -> bool

pub fn region(&self) -> &Region

pub fn check_before_tick(&self, cfg: &Config) -> CheckTickResult

Check whether the peer can be hibernated.

This should be used with check_after_tick to get a correct conclusion.

pub fn check_after_tick(&self, state: GroupState, res: CheckTickResult) -> bool

pub fn ping(&mut self)

Pings if followers are still connected.

Leader needs to know exact progress of followers, and followers just need to know whether leader is still alive.

pub fn set_region(
&mut self,
host: &CoprocessorHost<impl KvEngine>,
reader: &mut ReadDelegate,
region: Region
)

Set the region of a peer.

This will update the region of the peer, caller must ensure the region has been preserved in a durable device.

pub fn peer_id(&self) -> u64

pub fn leader_id(&self) -> u64

pub fn is_leader(&self) -> bool

pub fn get_role(&self) -> StateRole

pub fn get_store(&self) -> &PeerStorage<EK, ER>

pub fn mut_store(&mut self) -> &mut PeerStorage<EK, ER>

pub fn is_applying_snapshot(&self) -> bool

pub fn has_pending_snapshot(&self) -> bool

Returns true if the raft group has replicated a snapshot but not committed it yet.

pub fn get_pending_snapshot(&self) -> Option<&Snapshot>

fn add_ready_metric(&self, ready: &Ready, metrics: &mut RaftReadyMetrics)

fn send<T, I>(
&mut self,
trans: &mut T,
msgs: I,
metrics: &mut RaftMessageMetrics
) where
T: Transport,
I: IntoIterator<Item = Message>,

pub fn step<T, C>(
&mut self,
ctx: &mut PollContext<EK, ER, T, C>,
m: Message
) -> Result<()>

Steps the raft message.

pub fn check_peers(&mut self)

Checks and updates peer_heartbeats for the peer.

pub fn collect_down_peers(&mut self, max_duration: Duration) -> Vec<PeerStats>

Collects all down peers.

pub fn collect_pending_peers<T, C>(
&mut self,
ctx: &PollContext<EK, ER, T, C>
) -> Vec<Peer>

Collects all pending peers and update peers_start_pending_time.

pub fn any_new_peer_catch_up(&mut self, peer_id: u64) -> bool

Returns true if any peer recover from connectivity problem.

A peer can become pending or down if it has not responded for a long time. If it becomes normal again, PD need to be notified.

pub fn check_stale_state<T, C>(
&mut self,
ctx: &mut PollContext<EK, ER, T, C>
) -> StaleState

fn on_role_changed<T, C>(
&mut self,
ctx: &mut PollContext<EK, ER, T, C>,
ready: &Ready
)

pub fn ready_to_handle_pending_snap(&self) -> bool

fn ready_to_handle_read(&self) -> bool

fn ready_to_handle_unsafe_replica_read(&self, read_index: u64) -> bool

fn is_splitting(&self) -> bool

fn is_merging(&self) -> bool

fn is_merging_strict(&self) -> bool

pub fn ready_to_handle_request_snapshot(&mut self, request_index: u64) -> bool

pub fn can_early_apply(&self, term: u64, index: u64) -> bool

Whether a log can be applied before writing raft batch.

If TiKV crashes, it's possible apply index > commit index. If logs are still available in other nodes, it's possible to be recovered. But for singleton, logs are only available on single node, logs are gone forever.

Note we can't just check singleton. Because conf change takes effect on apply, so even there are two nodes, previous logs can still be committed by leader alone. Those logs can't be applied early. After introducing joint consensus, the node number can be undetermined. So here check whether log is persisted on disk instead.

Only apply existing logs has another benefit that we don't need to deal with snapshots that are older than apply index as apply index <= last index <= index of snapshot.

pub fn replication_mode_need_catch_up(&self) -> bool

Checks if leader needs to keep sending logs for follower.

In DrAutoSync mode, if leader goes to sleep before the region is sync, PD may wait longer time to reach sync state.

pub fn handle_raft_ready_append<T: Transport, C>(
&mut self,
ctx: &mut PollContext<EK, ER, T, C>
) -> Option<(Ready, InvokeContext)>

pub fn post_raft_ready_append<T: Transport, C>(
&mut self,
ctx: &mut PollContext<EK, ER, T, C>,
ready: &mut Ready,
invoke_ctx: InvokeContext
) -> Option<ApplySnapResult>

pub fn handle_raft_ready_apply<T, C>(
&mut self,
ctx: &mut PollContext<EK, ER, T, C>,
ready: &mut Ready,
invoke_ctx: &InvokeContext
)

pub fn handle_raft_ready_advance(&mut self, ready: Ready)

fn response_read<T, C>(
&self,
read: &mut ReadIndexRequest<EK::Snapshot>,
ctx: &mut PollContext<EK, ER, T, C>,
replica_read: bool
)

fn post_pending_read_index_on_replica<T, C>(
&mut self,
ctx: &mut PollContext<EK, ER, T, C>
)

Responses to the ready read index request on the replica, the replica is not a leader.

fn apply_reads<T, C>(
&mut self,
ctx: &mut PollContext<EK, ER, T, C>,
ready: &Ready
)

pub fn post_apply<T, C>(
&mut self,
ctx: &mut PollContext<EK, ER, T, C>,
apply_state: RaftApplyState,
applied_index_term: u64,
apply_metrics: &ApplyMetrics
) -> bool

pub fn post_split(&mut self)

fn maybe_renew_leader_lease<T, C>(
&mut self,
ts: Timespec,
ctx: &mut PollContext<EK, ER, T, C>,
progress: Option<ReadProgress>
)

Try to renew leader lease.

fn maybe_update_read_progress(
&self,
reader: &mut ReadDelegate,
progress: ReadProgress
)

pub fn maybe_campaign(&mut self, parent_is_leader: bool) -> bool

pub fn propose<T: Transport, C>(
&mut self,
ctx: &mut PollContext<EK, ER, T, C>,
cb: Callback<EK::Snapshot>,
req: RaftCmdRequest,
err_resp: RaftCmdResponse,
txn_extra: TxnExtra
) -> bool

Propose a request.

Return true means the request has been proposed successfully.

fn post_propose<T, C>(
&mut self,
poll_ctx: &mut PollContext<EK, ER, T, C>,
p: Proposal<EK::Snapshot>
)

fn check_conf_change<T, C>(
&mut self,
ctx: &mut PollContext<EK, ER, T, C>,
cmd: &RaftCmdRequest
) -> Result<()>

Validate the ConfChange request and check whether it's safe to propose the specified conf change request. It's safe iff at least the quorum of the Raft group is still healthy right after that conf change is applied. Define the total number of nodes in current Raft cluster to be total. To ensure the above safety, if the cmd is

A AddNode request Then at least '(total + 1)/2 + 1' nodes need to be up to date for now.

A RemoveNode request Then at least '(total - 1)/2 + 1' other nodes (the node about to be removed is excluded) need to be up to date for now. If 'allow_remove_leader' is false then the peer to be removed should not be the leader.

fn transfer_leader(&mut self, peer: &Peer)

fn pre_transfer_leader(&mut self, peer: &Peer) -> bool

fn ready_to_transfer_leader<T, C>(
&self,
ctx: &mut PollContext<EK, ER, T, C>,
index: u64,
peer: &Peer
) -> Option<&'static str>

fn read_local<T, C>(
&mut self,
ctx: &mut PollContext<EK, ER, T, C>,
req: RaftCmdRequest,
cb: Callback<EK::Snapshot>
)

fn pre_read_index(&self) -> Result<()>

pub fn has_unresolved_reads(&self) -> bool

pub fn retry_pending_reads(&mut self, cfg: &Config)

ReadIndex requests could be lost in network, so on followers commands could queue in pending_reads forever. Sending a new ReadIndex periodically can resolve this.

fn read_index<T: Transport, C>(
&mut self,
poll_ctx: &mut PollContext<EK, ER, T, C>,
req: RaftCmdRequest,
err_resp: RaftCmdResponse,
cb: Callback<EK::Snapshot>
) -> bool

pub fn get_min_progress(&self) -> Result<(u64, u64)>

Returns (minimal matched, minimal committed_index)

For now, it is only used in merge.

fn pre_propose_prepare_merge<T, C>(
&self,
ctx: &mut PollContext<EK, ER, T, C>,
req: &mut RaftCmdRequest
) -> Result<()>

fn pre_propose<T, C>(
&self,
poll_ctx: &mut PollContext<EK, ER, T, C>,
req: &mut RaftCmdRequest
) -> Result<ProposalContext>

fn propose_normal<T, C>(
&mut self,
poll_ctx: &mut PollContext<EK, ER, T, C>,
req: RaftCmdRequest
) -> Result<Either<u64, u64>>

Propose normal request to raft

Returns Ok(Either::Left(index)) means the proposal is proposed successfully and is located on index position. Ok(Either::Right(index)) means the proposal is rejected by CmdEpochChecker and the index is the position of the last conflict admin cmd.

fn execute_transfer_leader<T, C>(
&mut self,
ctx: &mut PollContext<EK, ER, T, C>,
msg: &Message
)

fn propose_transfer_leader<T, C>(
&mut self,
ctx: &mut PollContext<EK, ER, T, C>,
req: RaftCmdRequest,
cb: Callback<EK::Snapshot>
) -> bool

Return true to if the transfer leader request is accepted.

When transferring leadership begins, leader sends a pre-transfer to target follower first to ensures it's ready to become leader. After that the real transfer leader process begin.

pre_transfer_leader on leader: Leader will send a MsgTransferLeader to follower.

execute_transfer_leader on follower If follower passes all necessary checks, it will reply an ACK with type MsgTransferLeader and its promised persistent index.

execute_transfer_leader on leader: Leader checks if it's appropriate to transfer leadership. If it does, it calls raft transfer_leader API to do the remaining work.

See also: tikv/rfcs#37.

fn propose_conf_change<T, C>(
&mut self,
ctx: &mut PollContext<EK, ER, T, C>,
req: &RaftCmdRequest
) -> Result<Either<u64, u64>>

Returns Ok(Either::Left(index)) means the proposal is proposed successfully and is located on index position. Ok(Either::Right(index)) means the proposal is rejected by CmdEpochChecker and the index is the position of the last conflict admin cmd.

fn handle_read<T, C>(
&self,
ctx: &mut PollContext<EK, ER, T, C>,
req: RaftCmdRequest,
check_epoch: bool,
read_index: Option<u64>
) -> ReadResponse<EK::Snapshot>

pub fn term(&self) -> u64

pub fn stop(&mut self)

pub fn maybe_add_want_rollback_merge_peer(
&mut self,
peer_id: u64,
extra_msg: &ExtraMessage
)

pub fn add_want_rollback_merge_peer(&mut self, peer_id: u64)

impl<EK, ER> Peer<EK, ER> where
EK: KvEngine,
ER: RaftEngine,

pub fn insert_peer_cache(&mut self, peer: Peer)

pub fn remove_peer_from_cache(&mut self, peer_id: u64)

pub fn get_peer_from_cache(&self, peer_id: u64) -> Option<Peer>

fn region_replication_status(&mut self) -> Option<RegionReplicationStatus>

pub fn heartbeat_pd<T, C>(&mut self, ctx: &PollContext<EK, ER, T, C>)

fn send_raft_message<T: Transport>(&mut self, msg: Message, trans: &mut T)

pub fn bcast_wake_up_message<T: Transport, C>(
&self,
ctx: &mut PollContext<EK, ER, T, C>
)

pub fn bcast_check_stale_peer_message<T: Transport, C>(
&mut self,
ctx: &mut PollContext<EK, ER, T, C>
)

pub fn on_check_stale_peer_response(
&mut self,
check_conf_ver: u64,
check_peers: Vec<Peer>
)

pub fn send_want_rollback_merge<T: Transport, C>(
&self,
premerge_commit: u64,
ctx: &mut PollContext<EK, ER, T, C>
)