Trait raftstore::store::Snapshot

pub trait Snapshot<EK: KvEngine>: GenericSnapshot {
    fn build(
        &mut self, 
        engine: &EK, 
        kv_snap: &EK::Snapshot, 
        region: &Region, 
        snap_data: &mut RaftSnapshotData, 
        stat: &mut SnapshotStatistics
    ) -> RaftStoreResult<()>;

    fn apply(&mut self, options: ApplyOptions<EK>) -> Result<(), Error>;
}

Snapshot is a trait for snapshot. It’s used in these scenarios:

build local snapshot

read local snapshot and then replicate it to remote raftstores

receive snapshot from remote raftstore and write it to local storage

apply snapshot

snapshot gc

Required methods

fn build(
&mut self,
engine: &EK,
kv_snap: &EK::Snapshot,
region: &Region,
snap_data: &mut RaftSnapshotData,
stat: &mut SnapshotStatistics
) -> RaftStoreResult<()>

fn apply(&mut self, options: ApplyOptions<EK>) -> Result<(), Error>

Loading content...