Module raftstore::store::worker::region

Structs

PendingDeleteRanges
A structure records all ranges to be deleted with some delay. The delay is because there may be some coprocessor requests related to these ranges.

Runner

SnapContext

StalePeerInfo

Enums

Event
Region related timeout event

Task
Region related task

Constants

CLEANUP_MAX_REGION_COUNT

GENERATE_POOL_SIZE

PENDING_APPLY_CHECK_INTERVAL

STALE_PEER_CHECK_INTERVAL