Struct raftstore::store::fsm::apply::ApplyFsm

pub struct ApplyFsm<EK> where
    EK: KvEngine,  {
    delegate: ApplyDelegate<EK>,
    receiver: Receiver<Msg<EK>>,
    mailbox: Option<BasicMailbox<ApplyFsm<EK>>>,
}

Fields
delegate: ApplyDelegate<EK>receiver: Receiver<Msg<EK>>mailbox: Option<BasicMailbox<ApplyFsm<EK>>>
Implementations

impl<EK> ApplyFsm<EK> where
EK: KvEngine,

fn from_peer<ER: RaftEngine>(
peer: &Peer<EK, ER>
) -> (LooseBoundedSender<Msg<EK>>, Box<ApplyFsm<EK>>)

fn from_registration(
reg: Registration
) -> (LooseBoundedSender<Msg<EK>>, Box<ApplyFsm<EK>>)

fn handle_registration(&mut self, reg: Registration)

Handles peer registration. When a peer is created, it will register an apply delegate.

fn handle_apply<W: WriteBatch<EK>>(
&mut self,
apply_ctx: &mut ApplyContext<EK, W>,
apply: Apply<EK::Snapshot>
)

Handles apply tasks, and uses the apply delegate to handle the committed entries.

fn append_proposal(&mut self, props_drainer: Drain<'_, Proposal<EK::Snapshot>>)

Handles proposals, and appends the commands to the apply delegate.

fn destroy<W: WriteBatch<EK>>(&mut self, ctx: &mut ApplyContext<EK, W>)

fn handle_destroy<W: WriteBatch<EK>>(
&mut self,
ctx: &mut ApplyContext<EK, W>,
d: Destroy
)

Handles peer destroy. When a peer is destroyed, the corresponding apply delegate should be removed too.

fn resume_pending<W: WriteBatch<EK>>(&mut self, ctx: &mut ApplyContext<EK, W>)

fn logs_up_to_date_for_merge<W: WriteBatch<EK>>(
&mut self,
ctx: &mut ApplyContext<EK, W>,
catch_up_logs: CatchUpLogs
)

fn handle_snapshot<W: WriteBatch<EK>>(
&mut self,
apply_ctx: &mut ApplyContext<EK, W>,
snap_task: GenSnapTask
)

fn handle_change<W: WriteBatch<EK>>(
&mut self,
apply_ctx: &mut ApplyContext<EK, W>,
cmd: ChangeObserver,
region_epoch: RegionEpoch,
cb: Callback<EK::Snapshot>
)

fn handle_tasks<W: WriteBatch<EK>>(
&mut self,
apply_ctx: &mut ApplyContext<EK, W>,
msgs: &mut Vec<Msg<EK>>
)