Enum raftstore::store::msg::Callback

pub enum Callback<S: Snapshot> {
    None,
    Read(ReadCallback<S>),
    Write {
        cb: WriteCallback,
        proposed_cb: Option<ExtCallback>,
    },
}

Variants of callbacks for Msg.

Read: a callback for read only requests including StatusRequest, GetRequest and SnapRequest

Write: a callback for write only requests including AdminRequest PutRequest, DeleteRequest and DeleteRangeRequest.

Variants

None

No callback.

Read(ReadCallback<S>)

Read callback.

Write

Write callback.

Fields of Write

cb: WriteCallbackproposed_cb: Option<ExtCallback>

proposed_cb is called after a request is proposed to the raft group successfully. It's used to notify the caller to move on early because it's very likely the request will be applied to the raftstore.

Implementations

impl<S> Callback<S> where
S: Snapshot,

pub fn write(cb: WriteCallback) -> Self

pub fn write_ext(cb: WriteCallback, proposed_cb: Option<ExtCallback>) -> Self

pub fn invoke_with_response(self, resp: RaftCmdResponse)

pub fn invoke_proposed(&mut self)

pub fn invoke_read(self, args: ReadResponse<S>)

pub fn is_none(&self) -> bool